requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
coverage==4.5.3
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
PyJWT==1.7.1
pytest==4.4.1
pytest-cov==2.6.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
